ai-services/visual-generation/cropping.py
def crop_image(image_path: str) -> str:
    # TODO: Implement image cropping logic
    print(f"Cropping image at: {image_path}")
    return "/path/to/cropped_image.jpg"


# Reference Implementation
"""
from PIL import Image

def crop_image(image_path: str, aspect_ratio: tuple = (9, 16)) -> str:
    image = Image.open(image_path)
    width, height = image.size

    target_aspect = aspect_ratio[0] / aspect_ratio[1]
    current_aspect = width / height

    if current_aspect > target_aspect:
        # Crop the width
        new_width = int(target_aspect * height)
        left = (width - new_width) / 2
        top = 0
        right = (width + new_width) / 2
        bottom = height
    else:
        # Crop the height
        new_height = int(width / target_aspect)
        left = 0
        top = (height - new_height) / 2
        right = width
        bottom = (height + new_height) / 2

    cropped_image = image.crop((left, top, right, bottom))
    
    cropped_image_path = image_path.replace(".jpg", "_cropped.jpg")
    cropped_image.save(cropped_image_path)

    return cropped_image_path
"""
