ai-services/requirements.txt
fastapi
uvicorn
pydantic
openai
requests
Pillow
boto3
